Added MIT Liscence file
Hi Andrew! The work you did with this package was amazing. I'm Nilesh - a college freshman at NYU. I wanted to use this package in an application I'm trying to build which requires financial data. I'd prefer to be on the right side of copyright law, so I was hoping you could include this license file w/ your code. Thank you so much for your time =)